backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.5
python-dotenv==1.0.1
cryptography==43.0.1
libzim>=3.5.0
psycopg2-binary>=2.9
pymysql>=1.0
